==> packages/pimms-learn/pimms_learn-0.3.0.tar.gz/pimms_learn-0.3.0/project/02_2_join_configs.py.ipynb <==
# Join aggregated configs from each model

import pandas as pd

filepaths_in = snakemake.input
filepath_out = snakemake.output[0]
filepaths_in

filepath_out

## Example

- first file

pd.read_csv(filepaths_in[0], index_col=0)

## Load all model configs

configs = pd.concat([
    pd.read_csv(fname, index_col=0) for fname in filepaths_in
])
configs

## Dump combined to disk

_ = configs.to_csv(filepath_out)
filepath_out
